test: relocate shared skin fixture and delete leftover testing generator dir
The `testing` generator was removed during the API-coverage work but its
`images/` dir lingered because `testSheet.png` is still used as a shared
skin fixture. Move that fixture to a proper shared home
(`_common/fixtures/`), repoint all 6 consumers (1 src import in
testApiControls + 5 Playwright specs: heads, mini, block, item,
testApiControls), delete the unreferenced `thumbnail.png`, and remove the
now-empty `src/generators/testing/` directory.

No behaviour change. types:check + lint clean; the 5 fixture-consuming
specs pass (45 tests), including every skin/atlas upload case.
